+++ Assets/Scripts/Inventory/Model/EffectModifiers/BuffModifier.cs
@@ -6,7 +6,7 @@ using UnityEngine;
 public class BuffModifier : CharacterStatModifierSO
 {
     public int ID = 3;
-    public override void AffectCharacter(GameObject character, float val, BuffSO buff = null)
+    public override void AffectCharacter(GameObject character, int val, BuffSO buff = null)
     {
         if (character.tag == "Player" && buff != null)
         {
