+++ Assets/Scripts/Inventory/Model/EffectModifiers/CharacterStatModifierSO.cs
@@ -5,5 +5,5 @@ using UnityEngine;
 public abstract class CharacterStatModifierSO : ScriptableObject
 {
     //플레이어가 얻을 효과
-    public abstract void AffectCharacter(GameObject character, float val, BuffSO buff = null);
+    public abstract void AffectCharacter(GameObject character, int val, BuffSO buff = null);
 }
